Add inventory management tabs for overview, products, and services
- Implement OverviewTab for displaying overall inventory metrics and recent activity.
- Create ProductsTab for managing product listings with search and filter functionalities.
- Develop ServicesTab for managing services with similar features as ProductsTab.
- Integrate database interactions for loading, editing, and deleting products and services.
- Enhance UI with context menus, control panels, and advanced filtering options.
