+++ app/ui/pages/inventory/tabs/__init__.py
@@ -1,0 +1,5 @@
+# Import all tab classes for easy access
+from .overview_tab import OverviewTab
+from .products_tab import ProductsTab
+from .services_tab import ServicesTab
+from .inventory_status_tab import InventoryStatusTab
